+++ next-docs/tailwind.config.js
@@ -32,11 +32,6 @@ module.exports = {
         current: 'currentColor',
       },
       screens: {
-        sm: '640px',
-        md: '768px',
-        lg: '1024px',
-        xl: '1280px',
-        '2xl': '1536px',
         '3xl': '1800px',
       },
     },