+++ packages/popover/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Change Log
 
+## 9.15.0
+
+### Minor Changes
+
+- Version bump
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-assets@9.15.0
+  - @heathmont/moon-components@9.15.0
+  - @heathmont/moon-themes@9.15.0
+  - @heathmont/moon-utils@9.15.0
+
+## 9.14.3
+
+### Patch Changes
+
+- Bump version
+- Updated dependencies [undefined]
+  - @heathmont/moon-themes@9.14.3
+  - @heathmont/moon-assets@9.14.3
+  - @heathmont/moon-components@9.14.3
+  - @heathmont/moon-utils@9.14.3
+
 ## 9.14.2
 
 ### Patch Changes